chore(ci): skip cascade if [skip cascade] is found in the PR title (#476)
When there are multiple PRs in flight, using a var to disable the cascading PR that eventually runs the end to end tests (for instance because the change is about examples) is error prone.

Allowing each pull request to indivudually decide when it is relevant or not is more practical and fine grain.

## Testing

Using [an extra commit](https://code.forgejo.org/forgejo/runner/commit/1b8949ed3ada59ab0b1712fd5c3948f996757469) for debuging:

- https://code.forgejo.org/forgejo/runner/actions/runs/3328 skips the cascade on this PR
- https://code.forgejo.org/forgejo/runner/actions/runs/3326 tries to run the cascade on this PR  (and fails because it does not have secrets) because `[skip ZZ cascade]` is set in the title

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/476
Reviewed-by: Michael Kriese <michael.kriese@gmx.de>
Co-authored-by: Earl Warren <contact@earl-warren.org>
Co-committed-by: Earl Warren <contact@earl-warren.org>