fix: lxc-systemd: upgrade using a URL instead of a version [skip cascade] (#520)
It is common to install a version from a branch or a fork. Change the upgrade to be a URL instead of a version number for simplicity.

Fix a bug by which the lxc-helpers and the runner are upgraded but not the script itself.

Increase the service unit timeout when stopping the runner: it may take up to three hours.

Adapt the tests accordingly.

Fixes https://code.forgejo.org/forgejo/runner/issues/510

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/520
Reviewed-by: Michael Kriese <michael.kriese@gmx.de>
Co-authored-by: Earl Warren <contact@earl-warren.org>
Co-committed-by: Earl Warren <contact@earl-warren.org>