fix: ReportLog retry if noMore && more rows are to be sent
r.logOffset = ack
ack < r.logOffset+len(rows)

can be simplified as

len(rows) > 0

which is only false if there were no rows to send. It follows
that

if noMore && ack < r.logOffset+len(rows) {

is can be simplified as

if noMore {

is always true if rows were sent.

The intent was apparently to return on error and retry if only
part of the rows were sent but not all of them. To achieve that
the expression is replaced with:

if noMore && len(r.logRows) > 0 {