feat: allow inputs in runs-on (#1117)
Allows referencing the `inputs` context in `jobs.<job_id>.runs-on`. That is a prerequisite for enabling users to enter a label in the Forgejo UI during `workflow_dispatch`.

Example:

```yaml
name: test
on:
  workflow_dispatch:
    inputs:
      image:
        required: true
        type: string

jobs:
  test:
    runs-on: ${{ inputs.image }}
    steps:
      - run: echo "Running on ${{ inputs.image }}"
```

While Forgejo already has the necessary UI to enter the variable, it does not pass the context to the jobparser. That still has to be done.

Apart from enabling `inputs`, this PR expands the test coverage of the contexts supported by `runs-on`.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1117): <!--number 1117 --><!--line 0 --><!--description ZmVhdDogYWxsb3cgaW5wdXRzIGluIHJ1bnMtb24=-->feat: allow inputs in runs-on<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1117
Reviewed-by: Mathieu Fenniak <mfenniak@noreply.code.forgejo.org>
Co-authored-by: Andreas Ahlenstorf <andreas@ahlenstorf.ch>
Co-committed-by: Andreas Ahlenstorf <andreas@ahlenstorf.ch>