fix: improve logging & display of errors during workflow evaluation (#1165)
Fixes #1161.

The first commit improves the error message in the specific case when finding the `action.yml` file (and related) fails to a composite message.

The second commit introduces a new convention for forgejo-runner errors.  Currently some errors messages are being logged *only* to the runner's error logs because the errors are not marked with `raw_output` -- in other words, they're not the output of the step that is being executed, so they aren't logged *inside the step*, and at best they get logged in "Set up job" or "Complete job" rather than being close to where they occurred.  In the case that an error occurs when running a step, those are now logged within the step, and with the prefix "⚙️ [runner]:" to differentiate them from output from the step itself.

![image](/attachments/103ebdc8-5911-4a09-ad53-9e6a07cd63e7)

There are probably more locations that this new convention could be used to put log messages in a more useful/direct place, but this covers the obvious case of errors running a step.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1165): <!--number 1165 --><!--line 0 --><!--description Zml4OiBpbXByb3ZlIGxvZ2dpbmcgJiBkaXNwbGF5IG9mIGVycm9ycyBkdXJpbmcgd29ya2Zsb3cgZXZhbHVhdGlvbg==-->fix: improve logging & display of errors during workflow evaluation<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1165
Reviewed-by: limiting-factor <limiting-factor@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>