fix: 'failed to read action' errors when using relative workdir_parent (#1175)
When `workdir_parent` is a relative path, actions were failing with error:
```
⚙️ [runner]: failed to read action 'Checkout', no files found after reading paths: action.yml, action.yaml, Dockerfile
```

The cause of this is that worktree creation using `git -C relative/repodir worktree add relative/worktree` would actually create the `relative/worktree` inside the repo.  It seems that `git -C` to operate on a repo changes the working directory into that repo and evaluates relative paths from there.

Fixed by converting `workdir_parent` into an absolute path after the config file is read, and adding an error in the git operations if a relative path somehow makes its way through.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1175): <!--number 1175 --><!--line 0 --><!--description Zml4OiAnZmFpbGVkIHRvIHJlYWQgYWN0aW9uJyBlcnJvcnMgd2hlbiB1c2luZyByZWxhdGl2ZSB3b3JrZGlyX3BhcmVudA==-->fix: 'failed to read action' errors when using relative workdir_parent<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1175
Reviewed-by: limiting-factor <limiting-factor@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>