feat: evaluate variables in strategy.matrix during job parsing (#1188)
Although runner has tests that an evaluated matrix works in `act/runner` such as the [evalmatrix test](https://code.forgejo.org/forgejo/runner/src/commit/8e624ab12032894adeb7d0e55bb11e1eaceb79e3/act/runner/testdata/evalmatrix/push.yml), these workflows don't work in Forgejo because the `jobparser` component doesn't perform evaluation on the `strategy.matrix` node.

With this addition of matrix evaluation, it is possible to use matrix evaluation on input data as long as the data is known at execution time of the workflow.  This would be a prerequisite (and possibly qualify as a "workaround") for more the complicated dynamic job creation requested in https://code.forgejo.org/forgejo/forgejo-actions-feature-requests/issues/71.

For example, with a workflow such as this:
```yaml
name: semi-dynamic matrix

on:
  workflow_dispatch:
    inputs:
      platforms:
        description: 'platform list in JSON'
        required: true
        type: string

jobs:
  evalm:
    strategy:
      matrix:
        platforms: ${{ fromJson(inputs.platforms) }}
    runs-on: docker
    steps:
    - name: platform ${{ matrix.platform }}
      run: |
        echo $${{ matrix.platform }}
```

It's possible to have a dynamic matrix in a workflow dispatch (although this would make more sense with the API, not the UI here):
![image](/attachments/e76127f0-f73e-4e14-9b89-2acb53d65b40)

Resulting in multiple jobs being executed:
![image](/attachments/017b5c69-1417-4664-9521-cc4e8660ff95)

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1188): <!--number 1188 --><!--line 0 --><!--description ZmVhdDogZXZhbHVhdGUgdmFyaWFibGVzIGluIHN0cmF0ZWd5Lm1hdHJpeCBkdXJpbmcgam9iIHBhcnNpbmc=-->feat: evaluate variables in strategy.matrix during job parsing<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1188
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>