feat: run host child processes in groups (#1184)
This means process groups on POSIX and "job objects" on Windows. The
idea is that if the command gets cancelled (e.g. due to an Actions job
being cancelled), we need to terminate the entire process tree, not
just the root process.

This missing feature was the cause of forgejo/runner#1038. On POSIX
platforms, cancellation worked essentially by luck: programs tend to
terminate on SIGHUP (that being the default action for that signal), so
when the root process was killed, all of its children would receive
SIGHUP due to losing their controlling TTY and terminate. However, a
program *could* of course handle this signal or detach from the TTY, in
which case cancelation would previously fail to terminate them.

This patch has been tested on both Windows and Linux on the Zig
project's CI runners. Jobs were able to start normally, and could be
cancelled immediately, with no orphaned processes.

Resolves: forgejo/runner#1038

---

I wasn't sure whether to put this here or under windows/runner; the logic
affects all targets, but it's most significant on Windows. Let me know if I
should re-open this over there.

In any case, this is essentially achieving the same thing as windows/runner#2.
However, I believe this is a better patch: the other one is LLM-written, and appears
buggy and quite overcomplicated; plus, unlike this commit, it does not address the
(admittedly much less problematic) issue on POSIX targets.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- features
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1184): <!--number 1184 --><!--line 0 --><!--description ZmVhdDogcnVuIGhvc3QgY2hpbGQgcHJvY2Vzc2VzIGluIGdyb3Vwcw==-->feat: run host child processes in groups<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1184
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Matthew Lugg <mlugg@mlugg.co.uk>
Co-committed-by: Matthew Lugg <mlugg@mlugg.co.uk>