test: fix execution of TestCancelLongRunningCommand on NixOS (#1203)
Test from #1184 fails to execute in NixOS due to the `PATH` environment variable being uninitialized, so all commands used in the `evil.sh` script fail.  This is likely because there's no default behaviour of referencing `/bin` and `/usr/bin` in the NixOS shells, since those paths are only placeholders for `sh` and `env` on NixOS.

Added output from the `evil.sh` subcommand to the testing log when the `Eventually` fails as a diagnostic.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- other
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1203): <!--number 1203 --><!--line 0 --><!--description dGVzdDogZml4IGV4ZWN1dGlvbiBvZiBUZXN0Q2FuY2VsTG9uZ1J1bm5pbmdDb21tYW5kIG9uIE5peE9T-->test: fix execution of TestCancelLongRunningCommand on NixOS<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1203
Reviewed-by: earl-warren <earl-warren@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>