fix: double protocol prefix in reusable workflow clone URLs (#1213)
Tweaked version of https://code.forgejo.org/forgejo/runner/pulls/1212 with the normalization done outside the `RemoteReusableWorkflowWithHost` data structure, and an added automated test for regression prevent.  Credit to @syncstack for discovery and initial fix.

Fixes regression from #1209 in the parsing of relative remote reusable workflow URLs, if the runner config has calculated a `GitHubInstance` that is a URL and not a hostname.  The `GitHubInstance` value is used throughout the runner with flexibility for both options, but its usage in this way was missed in #1209.

<!--start release-notes-assistant-->
<!--URL:https://code.forgejo.org/forgejo/runner-->
- bug fixes
  - [PR](https://code.forgejo.org/forgejo/runner/pulls/1213): <!--number 1213 --><!--line 0 --><!--description Zml4OiBkb3VibGUgcHJvdG9jb2wgcHJlZml4IGluIHJldXNhYmxlIHdvcmtmbG93IGNsb25lIFVSTHM=-->fix: double protocol prefix in reusable workflow clone URLs<!--description-->
<!--end release-notes-assistant-->

Reviewed-on: https://code.forgejo.org/forgejo/runner/pulls/1213
Reviewed-by: limiting-factor <limiting-factor@noreply.code.forgejo.org>
Co-authored-by: Mathieu Fenniak <mathieu@fenniak.net>
Co-committed-by: Mathieu Fenniak <mathieu@fenniak.net>